Add seqspec for TAP-seq enhancer screen experiments from Schraivogel et al., 2020.
These experiments were performed usign the CROP-seq vector for guide delivery. The polyadenylated CROP-seq transcripts are captured as part of the targeted transcriptome readout and are contained in the R1 and R2 fastq files. The fastq files for these experiment are were already demultiplexed by the sequencing core facility, therefore no sample barcode is included in the seqspec.